pyproject.toml
[tool.poetry]
name = "docqa-stream"
version = "0.1.0"
description = ""
authors = ["ma2za <mazzapaolo2019@gmail.com>"]
readme = "README.md"
packages = [{ include = "src/docqa_stream" }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.101.1"
uvicorn = "^0.23.2"
unstructured = { extras = ["pdf"], version = "^0.10.5" }
weaviate-client = "^3.23.0"
langchain = "^0.0.272"
openai = "^0.27.9"
httpx = "^0.24.1"
pytest = "^7.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
